requirements.txt
playwright
openai
google-generativeai
python-dotenv
